+++ script/gulp/utils/get-version-from-argv.js
@@ -1,8 +1,12 @@
 'use strict'
 
-module.exports = () => {
+const semver = require('semver')
+
+module.exports = (defaultVersion = null) => {
   const arg = process.argv[process.argv.length - 1] || ''
   const version = arg.substring(2)
 
-  return version
+  return semver.valid(version) !== null
+    ? version
+    : defaultVersion
 }