+++ src/app.js
@@ -4,8 +4,8 @@ import './styles/index.sass'
 // import scripts
 import './scripts/index.js'
 
-// import icons
-require.context('./icons/', true, /\.svg$/)
+// import icon sprites
+require.context('./icons/sprites/', true, /\.svg$/)
 
 // import fonts
 require.context('./fonts/', true, /\.woff$/)