+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+## [1.3.1](https://github.com/herokuro/website/compare/v1.3.0...v1.3.1) (2019-09-21)
+
+
+### Bug Fixes
+
+* **webpack:** fix warning and error hints and large asset sizes ([c0c11e1](https://github.com/herokuro/website/commit/c0c11e1))
+
+# [1.3.0](https://github.com/herokuro/website/compare/v1.2.1...v1.3.0) (2019-09-21)
+
+
+### Bug Fixes
+
+* **jquery:** use jquery.slim instead of the regular version ([68e4fa9](https://github.com/herokuro/website/commit/68e4fa9))
+
+
+### Features
+
+* **metadata:** add og and twitter metadata with thumbnail images ([2d5fa9e](https://github.com/herokuro/website/commit/2d5fa9e))
+
+## [1.2.1](https://github.com/herokuro/website/compare/v1.2.0...v1.2.1) (2019-09-21)
+
+
+### Bug Fixes
+
+* **metadata:** fix website title and description metadata ([f4654dd](https://github.com/herokuro/website/commit/f4654dd))
+
+# [1.2.0](https://github.com/herokuro/website/compare/v1.1.0...v1.2.0) (2019-09-21)
+
+
+### Bug Fixes
+
+* **style:** add eslint rule exception to allow terser configs ([b420f2f](https://github.com/herokuro/website/commit/b420f2f))
+
+
+### Features
+
+* **metadata:** add favicon generation with data and versioning ([ff1f840](https://github.com/herokuro/website/commit/ff1f840))
+
+# [1.1.0](https://github.com/herokuro/website/compare/v1.0.2...v1.1.0) (2019-09-19)
+
+
+### Features
+
+* **website:** add social icons and images ([eed9ef6](https://github.com/herokuro/website/commit/eed9ef6))
+
 ## [1.0.2](https://github.com/herokuro/website/compare/v1.0.1...v1.0.2) (2019-09-17)